finished tutorial series
- finished tidying code

# same problem still persists
