+++ necosystem/Assets/scripts/MapPreview.cs
@@ -1,0 +1,88 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MapPreview : MonoBehaviour
+{
+    public Renderer textureRender;
+    public MeshFilter meshFilter;
+    public MeshRenderer meshRenderer;
+
+    public enum DrawMode { NoiseMap, Mesh, FalloffMap };
+    public DrawMode drawMode;
+
+    public MeshSettings meshSettings;
+    public HeightMapSettings heightMapSettings;
+    public textureData textureData;
+
+    public Material terrainMaterial;
+
+    [Range(0, MeshSettings.numSupportedLODs - 1)]
+    public int editorPreviewLOD;
+    public bool autoUpdate;
+
+
+    public void DrawMapInEditor()
+    {
+        textureData.ApplyToMaterial(terrainMaterial);
+        textureData.UpdateMeshHeights(terrainMaterial, heightMapSettings.minHeight, heightMapSettings.maxHeight);
+
+        HeightMap heightMap = heightMapGenerator.generateHeightMap(meshSettings.numVertsPerLine, meshSettings.numVertsPerLine, heightMapSettings, Vector2.zero);
+
+        if (drawMode == DrawMode.NoiseMap)
+            DrawTexture(textureGenerator.TextureFromHeightMap(heightMap));
+        else if (drawMode == DrawMode.Mesh)
+            DrawMesh(meshGenerator.GenerateTerrainMesh(heightMap.values, meshSettings, editorPreviewLOD));
+        else if (drawMode == DrawMode.FalloffMap)
+            DrawTexture(textureGenerator.TextureFromHeightMap(new HeightMap(falloffGenerator.generateFalloffMap(meshSettings.numVertsPerLine),0,1)));
+    }
+
+    public void DrawTexture(Texture2D texture)
+    {
+        textureRender.sharedMaterial.mainTexture = texture; //apply texture to plane in scene view
+        textureRender.transform.localScale = new Vector3(texture.width, 1, texture.height) / 4f; //make plane size scale to texture size
+
+        textureRender.gameObject.SetActive(true);
+        meshFilter.gameObject.SetActive(false);
+    }
+
+    public void DrawMesh(MeshData meshData)
+    {
+        meshFilter.sharedMesh = meshData.CreateMesh();
+
+        textureRender.gameObject.SetActive(false);
+        meshFilter.gameObject.SetActive(true);
+    }
+
+    void OnValuesUpdated()
+    {
+        if (!Application.isPlaying)
+        {
+            DrawMapInEditor();
+        }
+    }
+
+    void OnTextureValuesUpdated()
+    {
+        textureData.ApplyToMaterial(terrainMaterial);
+    }
+
+    private void OnValidate() //called automatically when a variable is changed in inspector
+    {
+        if (meshSettings != null)
+        {
+            meshSettings.OnValuesUpdated -= OnValuesUpdated;
+            meshSettings.OnValuesUpdated += OnValuesUpdated;
+        }
+        if (heightMapSettings != null)
+        {
+            heightMapSettings.OnValuesUpdated -= OnValuesUpdated;
+            heightMapSettings.OnValuesUpdated += OnValuesUpdated;
+        }
+        if (textureData != null)
+        {
+            textureData.OnValuesUpdated -= OnTextureValuesUpdated;
+            textureData.OnValuesUpdated += OnTextureValuesUpdated;
+        }
+    }
+}
