+++ necosystem/Assets/scripts/MapPreview.cs.meta
