--- necosystem/Assets/scripts/mapGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ccf49b97f562ba042b5650ef8f1823c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
