+++ necosystem/Assets/scripts/textureGenerator.cs
@@ -14,17 +14,17 @@ public static class textureGenerator
         return texture;
     }
 
-    public static Texture2D TextureFromHeightMap(float[,] heightMap)
+    public static Texture2D TextureFromHeightMap(HeightMap heightMap)
     {
-        int width = heightMap.GetLength(0);
-        int height = heightMap.GetLength(1); //retrieve x and y values from heightMap array
+        int width = heightMap.values.GetLength(0);
+        int height = heightMap.values.GetLength(1); //retrieve x and y values from heightMap array
 
 
         Color[] colourMap = new Color[width * height];
         for (int y = 0; y < height; y++)
             for (int x = 0; x < width; x++) //for every x and y position
             {
-                colourMap[y * width + x] = Color.Lerp(Color.black, Color.white, heightMap[x, y]); //every row of pixels = a colour between white and black
+                colourMap[y * width + x] = Color.Lerp(Color.black, Color.white, Mathf.InverseLerp(heightMap.minValue, heightMap.maxValue, heightMap.values[x, y])); //every row of pixels = a colour between white and black
             }
         
         return TextureFromColourMap(colourMap, width, height);
